\name{IPHM-package}
\alias{IPHM-package}
\alias{IPHM}
\docType{package}
\title{
Informative Priors Hierarchical Model
IPHM
}
\description{
Utilizing existing data can be carried out in a straightforward fashion under the Bayesian framework in which the repository of historical data can be exploited to build informative priors and used in new data analysis. In this work, using microarray data, we investigate the feasibility and effectiveness of deriving informative priors from historical data and using them in the problem of detecting differentially expressed genes. 
}
\details{
\tabular{ll}{
Package: \tab IPHM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-19\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ben Li

Maintainer: Ben Li<ben.li@emory.edu>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
